Fix various compiler warnings
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/anatomist/anatomist-free/branches/bug_fix@69896 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/lib/window3D/transformer.cc
@@ -312,7 +312,7 @@ bool TransformerActionData::redoable()
     }
 
     return (_maintrans->motionHistorySize() > 0 &&
-            _maintrans->motionHistoryIndex() < _maintrans->motionHistorySize() - 1);
+            static_cast<size_t>(_maintrans->motionHistoryIndex()) < _maintrans->motionHistorySize() - 1);
 }
 
 void TransformerActionData::resetTransform()
